// src/lib/schema.ts
export const CreateStudentSchema = z.object({
  name: z.string().min(1, "Name is required"),
  gender: z.nativeEnum(Gender),

  address: z.string().min(1, "Address is required"),
  collegeName: z.string().min(1, "College name is required"),
  branchName: z.string().min(1, "Branch name is required"),

  DOB: z.string().min(1, "Date of birth is required"),

  mobile: z.string().min(10, "Mobile number must be valid"),
  alternateMobile: z.string().optional().or(z.literal("")),

  email: z.string().email("Invalid email"),

  fatherName: z.string().optional().or(z.literal("")),
  fatherContact: z.string().optional().or(z.literal("")),
  fatherOccupation: z.string().optional().or(z.literal("")),

  motherName: z.string().optional().or(z.literal("")),
  motherContact: z.string().optional().or(z.literal("")),
  motherOccupation: z.string().optional().or(z.literal("")),

  admissionDate: z.string().min(1, "Admission date is required"),

  status: z.nativeEnum(StudentStatus),

  batchId: z.string().min(1, "Batch is required"),
});

export type CreateStudentFormValues = z.infer<typeof CreateStudentSchema>;